Incremental Progess

+++ src/util/alias.rs
@@ -1,2 +1,2 @@
 pub type ID = String;
-pub type Dist = f64;
+pub type Dist = u64;
